refactor resource helpers
